rename: unit test file 하나의 폴더에서 관리하도록 변경
