+++ cypress/component/CommonCheckBox.cy.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CommonCheckBox } from './CommonCheckBox';
+import { CommonCheckBox } from '../../src/components/common/CommonCheckBox';
 
 describe('<CommonCheckBox />', () => {
   it('체크박스의 label이 제대로 렌더링되는지 확인', () => {
