+++ cypress/component/LoadingComponent.cy.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LoadingComponent } from './LoadingComponent';
+import { LoadingComponent } from '../../src/components/common/LoadingComponent';
 
 describe('<LoadingComponent />', () => {
   it('애니메이션 적용 여부', () => {
